# docs/api/sancho-openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Sancho.API | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5099/"
    }
  ],
  "paths": {
...
    "/api/identity/users/{userId}": {
      "get": {
        "tags": [
          "IdentityEndpoints"
        ],
        "operationId": "GetIdentityUserDetail",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/identity/users/{userId}/org-role": {
      "post": {
        "tags": [
          "IdentityEndpoints"
        ],
        "operationId": "AssignOrgRole",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignOrgRoleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "IdentityEndpoints"
        ],
        "operationId": "RevokeOrgRole",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AssignOrgRoleRequest": {
        "type": "object"
      },
...
    }
  },
...
}